Small refactoring done
