+++ FE/angular/src/app/components/movie/movie.component.html
@@ -1,6 +1,6 @@
 <app-go-back></app-go-back>
 
-<ul *ngIf="movie" class="movie">
+<ul *ngIf="movie$ | async as movie" class="movie">
   <li class="grid grid-bleed">
     <div class="image col-3">
       <img [src]="movie.Poster" [alt]="movie.Title + '-poster'" />
