+++ FE/angular/src/app/components/movie/movie.component.ts
@@ -1,14 +1,15 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Subscription, tap } from 'rxjs';
-import { DataService, MovieComplete } from '../../services/data.service';
+import { Observable, Subscription } from 'rxjs';
+import { MovieComplete } from 'src/app/models/data.model';
+import { DataService } from '../../services/data.service';
 
 @Component({
   selector: 'app-movie',
   templateUrl: './movie.component.html'
 })
 export class MovieComponent implements OnDestroy, OnInit {
-  public movie: MovieComplete;
+  public movie$: Observable<MovieComplete>;
   private movieSubscription: Subscription;
   public showError: boolean = false;
 
@@ -17,15 +18,11 @@ export class MovieComponent implements OnDestroy, OnInit {
   public ngOnInit() {
     this.activatedRoute.params.subscribe(({ id }) => {
       if (id) {
-        this.getMovieDetails(id);
+        this.movie$ = this.dataService.getMovie(id);
       } else {
         this.showError = true;
       }
     });
-  }
-
-  private getMovieDetails(movieId: string) {
-    this.movieSubscription = this.dataService.getMovie(movieId).subscribe((data) => (this.movie = data));
   }
 
   public ngOnDestroy(): void {
